Move contract related code to contract folder
