+++ src/contract/outcome.ts
