+++ src/contract/state.ts
