+++ test/ERC20AssetHolder/deposit.test.ts
@@ -10,8 +10,8 @@ import {
   newTransferEvent,
   sendTransaction,
 } from '../test-helpers';
-import {Channel, getChannelId} from '../../src/channel';
-import {createDepositTransaction} from '../../src/transaction-creators/erc20-asset-holder';
+import {Channel, getChannelId} from '../../src/contract/channel';
+import {createDepositTransaction} from '../../src/contract/transaction-creators/erc20-asset-holder';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
