+++ test/ForceMove/concludeFromOpen.test.ts
@@ -15,10 +15,10 @@ import {
   sendTransaction,
 } from '../test-helpers';
 import {HashZero, AddressZero} from 'ethers/constants';
-import {Outcome, hashOutcome} from '../../src/outcome';
-import {Channel, getChannelId} from '../../src/channel';
-import {State} from '../../src/state';
-import {createConcludeFromOpenTransaction} from '../../src/transaction-creators/force-move';
+import {Outcome, hashOutcome} from '../../src/contract/outcome';
+import {Channel, getChannelId} from '../../src/contract/channel';
+import {State} from '../../src/contract/state';
+import {createConcludeFromOpenTransaction} from '../../src/contract/transaction-creators/force-move';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
