+++ test/test-helpers.ts
@@ -9,7 +9,7 @@ import {
 } from 'ethers/utils';
 import {AddressZero} from 'ethers/constants';
 
-import {hashChannelStorage} from '../src/channel-storage';
+import {hashChannelStorage} from '../src/contract/channel-storage';
 import {
   Outcome,
   encodeAllocation,
@@ -17,8 +17,8 @@ import {
   encodeGuarantee,
   Guarantee,
   Allocation,
-} from '../src/outcome';
-import {State, hashState} from '../src/state';
+} from '../src/contract/outcome';
+import {State, hashState} from '../src/contract/state';
 import {TransactionRequest, TransactionReceipt} from 'ethers/providers';
 import {toHex} from '../src/hex-utils';
